backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
ldap3==2.9.1
pydantic-settings==2.10.1
python-multipart==0.0.20
